updated conversations styling and reorganized components
